% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{calcul_par_genre}
\alias{calcul_par_genre}
\title{Calculer le rating moyen par genre}
\usage{
calcul_par_genre(df, stat = "mean")
}
\arguments{
\item{df}{Un data.frame avec colonnes Genres et Rating.}

\item{stat}{La statistique : "mean", "max" ou "median".}
}
\value{
Un data.frame avec le rating par genre.
}
\description{
Calculer le rating moyen par genre
}
\examples{
calcul_par_genre(df_games, stat = "mean")
}
