% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_games}
\alias{df_games}
\title{Jeux vidéo populaires (2000-2023)}
\format{
Un data.frame avec 984 lignes et 10 variables :
\describe{
\item{Title}{Titre du jeu}
\item{Release Date}{Date de sortie}
\item{Rating}{Note moyenne}
\item{Times Listed}{Nombre de fois listé}
\item{Number of Reviews}{Nombre de critiques}
\item{Genres}{Genres du jeu}
\item{Plays}{Nombre de parties jouées}
\item{Playing}{Joueurs actuellement actifs}
\item{Backlogs}{Jeux en backlog}
\item{Wishlist}{Jeux en wishlist}
}
}
\source{
\url{https://www.kaggle.com/datasets/arnabchaki/popular-video-games-1980-2023}
}
\usage{
df_games
}
\description{
Dataset contenant des informations sur 984 jeux vidéo sortis après 2000.
}
\keyword{datasets}
