% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{filtrer_genre}
\alias{filtrer_genre}
\title{Filtrer les jeux par genre}
\usage{
filtrer_genre(df, genre)
}
\arguments{
\item{df}{Un data.frame contenant une colonne Genres.}

\item{genre}{Un caractère, le genre à rechercher (ex: "RPG").}
}
\value{
Un data.frame filtré.
}
\description{
Filtrer les jeux par genre
}
\examples{
filtrer_genre(df_games, "RPG")
}
