% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{plot_par_genre}
\alias{plot_par_genre}
\title{Visualiser le rating par genre}
\usage{
plot_par_genre(df, stat = "mean", n_genres = 10)
}
\arguments{
\item{df}{Un data.frame avec colonnes Genres et Rating.}

\item{stat}{La statistique à afficher : "mean", "max" ou "median".}

\item{n_genres}{Le nombre de genres à afficher (défaut : 10).}
}
\value{
Un objet ggplot2.
}
\description{
Visualiser le rating par genre
}
\examples{
plot_par_genre(df_games)
}
